// src/app/store/types/productsActionTypes.ts
export enum ActionProductsTypes {
  GET_PRODUCTS = '[Products/API] Get Products Request',
  GET_PRODUCTS_SUCCESS = '[Products/API] Get Products Success',
  GET_PRODUCTS_FAILURE = '[Products/API] Get Products Failure',

  ADD_PRODUCT = '[Products/API] Add product Request',
  ADD_PRODUCT_SUCCESS = '[Products/API] Add product Success',
  ADD_PRODUCT_FAILURE = '[Products/API] Add product Failure',

  ADD_PRODUCT_CLEAR = '[Products/API] Add product Clear',

  UPDATE_PRODUCT = '[Products/API] Update product Request',
  UPDATE_PRODUCT_SUCCESS = '[Products/API] Update product Success',
  UPDATE_PRODUCT_FAILURE = '[Products/API] Update product Failure',
}